Resize white bar to 92px width + 11px borders
I computed the height of the white bar to reach to a middle position
between the red bar (top = 320px) and the blue bar (top = 206px).

I had set the height of the blue bar to be twice as high as its width,
without borders: 92*2 = 184px. Adding the borders of 11px, the blue
bar reaches 206px.

The middle position between top of blue bar and top of red bar is:
(206 + 320) / 2 = 526 / 2 = 263px.

The height of the white bar without borders is thus:
263 - 2*11 = 263 - 22 = 241px.