Center text vertically and horizontally
I have added an intermediate with a white rectangle
at the dimensions of the text bounding box to find out
the centered position of the text before moving it.

Note: I have left the layers with the previous text position
in this version. I will remove them in next commit.